feat(GIFT-13654): update api tests
